requirements.txt
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
Faker>=25.0
python-dotenv>=1.0
tabulate>=0.9
